#95 Create the nex class of test AbstractAgent.java TestAbstractAgent.java
